# platform/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorEnvelopeResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelopeResponse",
        "type": "object"
      },
...
      "SubmissionPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SubmissionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "SubmissionPage",
        "type": "object"
      },
      "SubmissionResponse": {
        "additionalProperties": false,
        "properties": {
          "batch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "upload_id": {
            "title": "Upload Id",
            "type": "string"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "upload_id",
          "label",
          "batch",
          "source",
          "filename",
          "version",
          "submitted_at",
          "verified_at"
        ],
        "title": "SubmissionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Anonymous trusted-intranet crash analysis control plane. There are intentionally no login, RBAC, or DELETE endpoints.",
    "title": "Crash-Cap API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v3/workspaces/{workspace_id}/occurrences/{occurrence_id}/submissions": {
      "get": {
        "operationId": "list_submissions_api_v3_workspaces__workspace_id__occurrences__occurrence_id__submissions_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "List Submissions"
      }
    },
...
  }
}